requirements.txt
numpy
pandas
streamlit
seaborn
requests
pillow
plotly
scikit-learn